geoip processor support via geoip filter + ls-core database manger
First: I don't like this, and maintaining it will be a nightmare. The LS-core
provided geoip database manager is VERY tightly-coupled with the GeoIp filter.

While this change attempts to make a useful API boundary with a new in-project
support adapter so that the caller of the managed geoip database does not
need to be concerned with implementation details of subscribing to changes from
the manager, the existing tight-coupling is inescapable and results in a
runtime dependency on the GeoIp filter to satisfy LS core's need to use it to
validate the shape of the database that adter downloading it.

While I have some measure of confidencec that this will work, it is going to
be almost impossible to test in no small part because of the tight-coupling that
makes test coverage so necessary.